# AVX512+AVXVNNI GEMM implementation for quants using Q8_K for activations 