This is an alternative to #610 

The expectation is that it will significantly improve CPU-only prompt processing speed on "true" `AVX512` CPUs that support the  `AVX512VNNI, AVX512VL, AVX512BW,` and `AVX512DQ` extensions (e.g., Zen5). These extensions are also supported by Zen4 cores, but these are not "true" `AVX512` CPUs as there 512-bit instructions are performed as two 256-bit instructions. The main benefit compared to #610 is that on such CPUs performance is about the same as on main, unlike #610, where we get a 10-15% performance penalty.IQ2_XXS

The PR adds a new quantization type, `Q8_K_R16`, which is only used for quantizing activation tensors. `IQ1_S, IQ1_M, IQ2_XXS, IQ2_XS, IQ2_S, IQ3_XXS, IQ3_S, IQ4_XS, Q2_K, Q3_K, IQ2_KS, IQ2_K, IQ3_KS, IQ3_K, IQ4_KSS, IQ4_KS, IQ4_K, IQ5_KS, IQ5_K, IQ6_K` all use this new quantization type for GEMM when the batch size is `>= 32` and "fancy SIMD" are available (i.e., the above `AVX512` extensions are supported by the CPU).

If you have a CPU that supports `AVX512VNNI, AVX512VL, AVX512BW,` and `AVX512DQ` extensions (Ryzen 99XX, EPYCs of the 9005 series (a.k.a., Turin), recent Intel Xeon CPUs), please test and let me know if this PR improves CPU performance.

@ubergarm  Pinging you explicitly.