@ikawrakow 2025-08-21
@sousekd 

Thanks for testing. Short of downloading the exact same model and running the exact same command as you I'm absolutely not able to reproduce the assert. But I have pushed a change that hopefully fixes that.

Also, because there is some doubt if you are actually enabling the faster "fancy" SIMD path, I have added a hack that will tell us if it is enabled. If you pull the latest, rebuild, and run any command, you will in then output either
```
======================================= HAVE_FANCY_SIMD is defined
```
or
```
======================================= HAVE_FANCY_SIMD is NOT defined
```