@ikawrakow 2025-08-21
@ubergarm 

Thanks for testing! 

It looks like in your case `-rtr` is faster than no `-rtr`? This is interesting. On my 7950X, with this PR and the model you have used I get

### With -rtr

 |    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    1.153 |   443.88 |    5.179 |    24.72 |
|   512 |    128 |    512 |    1.194 |   428.89 |    5.263 |    24.32 |
|   512 |    128 |   1024 |    1.265 |   404.77 |    5.375 |    23.81 |
|   512 |    128 |   1536 |    1.348 |   379.75 |    5.484 |    23.34 |
|   512 |    128 |   2048 |    1.423 |   359.90 |    5.573 |    22.97 |
|   512 |    128 |   2560 |    1.497 |   342.03 |    5.666 |    22.59 |
|   512 |    128 |   3072 |    1.579 |   324.29 |    5.752 |    22.25 |
|   512 |    128 |   3584 |    1.686 |   303.63 |    5.845 |    21.90 |
|   512 |    128 |   4096 |    1.738 |   294.67 |    5.958 |    21.48 |

### Without -rtr

|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    1.008 |   508.03 |    4.855 |    26.37 |
|   512 |    128 |    512 |    1.046 |   489.32 |    5.102 |    25.09 |
|   512 |    128 |   1024 |    1.116 |   458.62 |    5.006 |    25.57 |
|   512 |    128 |   1536 |    1.194 |   428.88 |    5.098 |    25.11 |
|   512 |    128 |   2048 |    1.264 |   405.05 |    5.193 |    24.65 |
|   512 |    128 |   2560 |    1.336 |   383.25 |    5.263 |    24.32 |
|   512 |    128 |   3072 |    1.428 |   358.63 |    5.433 |    23.56 |
|   512 |    128 |   3584 |    1.514 |   338.27 |    5.551 |    23.06 |
|   512 |    128 |   4096 |    1.590 |   322.10 |    5.615 |    22.80 |

In any case, it would be useful to have the benchmark results for the Qwen3-14B model that you used in #610 to know if this PR is as good as #610