@saood06 2025-08-21
>But it never became painful enough so I sit down and do it.

Yep. Although I have occasionally hard coded 32 instead of ubatch/4, but the better solution would do TG differently (PP pass to build the KV, TG going backwards by removing from the cache until you are wherever you want to be to collect data).